+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - rename `HTTPSpec.Request.URL.parse/1` to `HTTPSpec.Request.URL.parse!/1`
 - add `HTTPSpec.Request.URL.parse/1` which returns ok/error tuple.
 - move all header related function to a new module `HTTPSpec.Header`
+- move all trailer related function to a new module `HTTPSpec.Trailer`
 
 ## 2.3.0